Create FreeArduFall
Display used: LCD 128X64
 3 pull-up buttons
 4 pairs of LDRs and LEDs
 Servo motor
Antenna clamp
Arduino Nano/UNO. 
The servo motor was powered by an external 5V power supply. Test object (Styrofoam ball with a radius of 5cm). 
PVC pipe with a diameter of 10cm (100mm) and a height of 1 meter. 
Distribution of the LDR-LED pairs along the pipe:
 the first pair was placed 5cm from the top entry of the pipe, and the others were placed every 10cm